Changes to close Issue #1
This issue was caused by the same root Slack API behavior that
contributed to Issue #5. The Slack API distinguishes between groups and
channels in its API calls, and any call made must use the appropriate
call for its type. Because Slack does not automatically format in-text
links to groups (as it does with channels), the usable ID for a group
must be found before it can be used as a target for any message.

Additionally, the `find_group` function is now decorated by a
memoization class to make calls to it less expensive. It should now keep
a running cache of previous arguments and their return values, to reduce
both API calls and execution time. Periodic clearing of this cache may
be implemented later to prevent overly-stale responses.