Implementation of `whenaway` module
New `whenaway.py` module should allow for better usability of this
library. Whenever a problem is proposed to a specific group, if this
module is enabled, the program will first check that at least one
member of the targeted team is not "away" in Slack. If no members are
active, the program will instead direct the problem proposal as widely
as possible, sending it to all admin channels instead of the specified
single channel. This will assure the highest exposure to admins. This
module can be disabled in `settings.json`.