Fixed misdirected denials when using whenaway module
The whenaway module is meant to direct problem proposals to the
defaulting channel whenever the recipients in the user's intended
channel are not active in Slack. This part of the behavior was working
correctly. However, when one of these problems was denied by the
defaulting channel, the resulting denial confirmation (which is
important user feedback) was delivered to the user's originally
intended channel. Because this proposal was never delivered to that
intended channel, there was no reason that it should ever see the
denial confirmation.

This has now been resolved, and denial confirmations for problems
that were redirected to the defaulting channel should now be properly
sent back to the defaulting channel.